rename metamask web wallet file
